o1vm/pickles: simplifying main file by extracting all constraints
